Moved BiomeProviderRegisterEvent to api package
